\name{makeCoCoCg}
\alias{makeCoCoCg}

\title{ Make a CoCoCg object }

\description{ Make a CoCoCg object. }

\usage{
makeCoCoCg(...)
}

\arguments{
  \item{\dots}{ \code{\link[CoCo]{make.coco}}. }
}

\details{ The distributed source code is only able to handle one concurrent
	  CoCo object. Use \code{\link[CoCo]{endCoCo}}. }

% \value{
%   ~Describe the value returned
%   If it is a LIST, use
%   \item{comp1 }{Description of 'comp1'}
%   \item{comp2 }{Description of 'comp2'}
%   ...
% }

% \references{ }

\author{ Jens Henrik Badsberg }

% \note{ }

% \section{ToDo}{....}

% \seealso{ \code{\link[CoCoCore]{make.coco}}. }

% \examples{ }

\keyword{ interface }% at least one, from doc/KEYWORDS
\keyword{ multivariate }% __ONLY ONE__ keyword per line
